refactor(workflows): extract definition routes into lib/routes/definitions (C2a)
Phase C2a of the workflows/index.ts split. Pulls the definition-management
route group out of the module-load populateWorkflowRoutes mutation into a
typed defineRoute[] module, now that handlers reach ctx via the accessor.

- lib/route-schemas.ts — the shared passthroughWf/errorResponseWf/
  htmlResponseWf/repairSkillBodyWf zod consts (one copy for every route group).
- lib/routes/definitions.ts — exports `definitionRoutes`: notification-channels
  list, /definitions GET/POST/PUT/DELETE + /:name GET, /skills/:name/repair,
  availability PATCH, /node-types. The user-owned-YAML CRUD helpers
  (getDefinitionsDir/findExistingUserDefinitionPath/writeUserDefinition) move
  with them — they had no callers outside this group.

definePlugin now spreads `[...definitionRoutes, ...workflowRoutes]`;
populateWorkflowRoutes is down to the instance + gate routes (C2b/C2c).
index.ts sheds ~10 now-unused imports and its local zod-const copies.

Pure behavior-preserving relocation — the route objects are byte-identical,
only their assembly site moved.

Verified: typecheck + eslint + workflows suite (420/0) + full suite (5124/0)
+ binary build + isolated-home boot smoke (GET /definitions, /node-types,
/notification-channels return 200; POST /definitions with a bad body → 400).
